requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
APScheduler==3.6.3
async-timeout==4.0.3
attrs==23.1.0
cachetools==4.2.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
multidict==6.0.4
mysql-connector-python==8.1.0
openai==0.27.9
protobuf==4.21.12
python-telegram-bot==13.7
pytz==2023.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tornado==6.3.3
tqdm==4.66.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
yarl==1.9.2
